+++ src/app/Http/routes.php
@@ -1,6 +1,6 @@
 <?php
 
-Route::group(['prefix' => 'admin', 'middleware' => 'auth'], function()
+Route::group(['prefix' => 'admin', 'middleware' => ['web', 'auth']], function()
 {
 
 	// Backup