+++ CHANGELOG.md
@@ -18,3 +18,10 @@ All Notable changes to `backupmanager` will be documented in this file
 
 ### Security
 - Nothing
+
+
+## 1.1 - 2016-03-15
+
+### Added
+- Updated to v3 of spatie's laravel-backup package.
+- Renamed everything to be part of Backpack instead of Dick.